app/requirements.txt
Pillow==10.4.0
Jinja2==3.1.4
PyYAML==6.0.2
Flask==3.0.3
requests==2.32.3
gunicorn==22.0.0
